se realiza HU-8
