+++ backend-v1/tests/reports/coverage-html/backend-v1/index.html
@@ -23,30 +23,30 @@
         <div class='clearfix'>
             
             <div class='fl pad1y space-right2'>
-                <span class="strong">14.89% </span>
+                <span class="strong">6.45% </span>
                 <span class="quiet">Statements</span>
-                <span class='fraction'>35/235</span>
+                <span class='fraction'>27/418</span>
             </div>
         
             
             <div class='fl pad1y space-right2'>
-                <span class="strong">2% </span>
+                <span class="strong">0.35% </span>
                 <span class="quiet">Branches</span>
-                <span class='fraction'>3/150</span>
+                <span class='fraction'>1/282</span>
             </div>
         
             
             <div class='fl pad1y space-right2'>
-                <span class="strong">4.76% </span>
+                <span class="strong">0% </span>
                 <span class="quiet">Functions</span>
-                <span class='fraction'>1/21</span>
+                <span class='fraction'>0/44</span>
             </div>
         
             
             <div class='fl pad1y space-right2'>
-                <span class="strong">15.62% </span>
+                <span class="strong">6.73% </span>
                 <span class="quiet">Lines</span>
-                <span class='fraction'>35/224</span>
+                <span class='fraction'>27/401</span>
             </div>
         
             
@@ -79,33 +79,18 @@
 </tr>
 </thead>
 <tbody><tr>
-	<td class="file high" data-value="firebase.js"><a href="firebase.js.html">firebase.js</a></td>
-	<td data-value="100" class="pic high">
-	<div class="chart"><div class="cover-fill cover-full" style="width: 100%"></div><div class="cover-empty" style="width: 0%"></div></div>
-	</td>
-	<td data-value="100" class="pct high">100%</td>
-	<td data-value="9" class="abs high">9/9</td>
-	<td data-value="100" class="pct high">100%</td>
-	<td data-value="2" class="abs high">2/2</td>
-	<td data-value="100" class="pct high">100%</td>
-	<td data-value="0" class="abs high">0/0</td>
-	<td data-value="100" class="pct high">100%</td>
-	<td data-value="9" class="abs high">9/9</td>
-	</tr>
-
-<tr>
 	<td class="file low" data-value="hybridServer.js"><a href="hybridServer.js.html">hybridServer.js</a></td>
-	<td data-value="11.5" class="pic low">
-	<div class="chart"><div class="cover-fill" style="width: 11%"></div><div class="cover-empty" style="width: 89%"></div></div>
+	<td data-value="6.45" class="pic low">
+	<div class="chart"><div class="cover-fill" style="width: 6%"></div><div class="cover-empty" style="width: 94%"></div></div>
 	</td>
-	<td data-value="11.5" class="pct low">11.5%</td>
-	<td data-value="226" class="abs low">26/226</td>
-	<td data-value="0.67" class="pct low">0.67%</td>
-	<td data-value="148" class="abs low">1/148</td>
-	<td data-value="4.76" class="pct low">4.76%</td>
-	<td data-value="21" class="abs low">1/21</td>
-	<td data-value="12.09" class="pct low">12.09%</td>
-	<td data-value="215" class="abs low">26/215</td>
+	<td data-value="6.45" class="pct low">6.45%</td>
+	<td data-value="418" class="abs low">27/418</td>
+	<td data-value="0.35" class="pct low">0.35%</td>
+	<td data-value="282" class="abs low">1/282</td>
+	<td data-value="0" class="pct low">0%</td>
+	<td data-value="44" class="abs low">0/44</td>
+	<td data-value="6.73" class="pct low">6.73%</td>
+	<td data-value="401" class="abs low">27/401</td>
 	</tr>
 
 </tbody>
@@ -116,7 +101,7 @@
             <div class='footer quiet pad2 space-top1 center small'>
                 Code coverage generated by
                 <a href="https://istanbul.js.org/" target="_blank" rel="noopener noreferrer">istanbul</a>
-                at 2025-10-11T01:53:09.221Z
+                at 2025-10-12T00:36:14.921Z
             </div>
         <script src="../prettify.js"></script>
         <script>
