+++ backend-v1/tests/reports/coverage-html/backend-v1/routes/ai.js.html
@@ -154,7 +154,7 @@ module.exports = router;
             <div class='footer quiet pad2 space-top1 center small'>
                 Code coverage generated by
                 <a href="https://istanbul.js.org/" target="_blank" rel="noopener noreferrer">istanbul</a>
-                at 2025-10-11T01:53:09.221Z
+                at 2025-10-12T00:36:14.921Z
             </div>
         <script src="../../prettify.js"></script>
         <script>
