+++ backend-v1/tests/reports/coverage-html/backend-v1/routes/index.html
@@ -23,30 +23,30 @@
         <div class='clearfix'>
             
             <div class='fl pad1y space-right2'>
-                <span class="strong">78.51% </span>
+                <span class="strong">39.64% </span>
                 <span class="quiet">Statements</span>
-                <span class='fraction'>106/135</span>
+                <span class='fraction'>67/169</span>
             </div>
         
             
             <div class='fl pad1y space-right2'>
-                <span class="strong">82.35% </span>
+                <span class="strong">0% </span>
                 <span class="quiet">Branches</span>
-                <span class='fraction'>42/51</span>
+                <span class='fraction'>0/71</span>
             </div>
         
             
             <div class='fl pad1y space-right2'>
-                <span class="strong">46.66% </span>
+                <span class="strong">0% </span>
                 <span class="quiet">Functions</span>
-                <span class='fraction'>7/15</span>
+                <span class='fraction'>0/15</span>
             </div>
         
             
             <div class='fl pad1y space-right2'>
-                <span class="strong">78.51% </span>
+                <span class="strong">40.6% </span>
                 <span class="quiet">Lines</span>
-                <span class='fraction'>106/135</span>
+                <span class='fraction'>67/165</span>
             </div>
         
             
@@ -61,7 +61,7 @@
             </div>
         </template>
     </div>
-    <div class='status-line medium'></div>
+    <div class='status-line low'></div>
     <div class="pad1">
 <table class="coverage-summary">
 <thead>
@@ -80,17 +80,17 @@
 </thead>
 <tbody><tr>
 	<td class="file low" data-value="adminAccessibility.js"><a href="adminAccessibility.js.html">adminAccessibility.js</a></td>
-	<td data-value="33.33" class="pic low">
-	<div class="chart"><div class="cover-fill" style="width: 33%"></div><div class="cover-empty" style="width: 67%"></div></div>
+	<td data-value="19.14" class="pic low">
+	<div class="chart"><div class="cover-fill" style="width: 19%"></div><div class="cover-empty" style="width: 81%"></div></div>
 	</td>
-	<td data-value="33.33" class="pct low">33.33%</td>
-	<td data-value="24" class="abs low">8/24</td>
+	<td data-value="19.14" class="pct low">19.14%</td>
+	<td data-value="47" class="abs low">9/47</td>
 	<td data-value="0" class="pct low">0%</td>
-	<td data-value="2" class="abs low">0/2</td>
+	<td data-value="22" class="abs low">0/22</td>
 	<td data-value="0" class="pct low">0%</td>
 	<td data-value="4" class="abs low">0/4</td>
-	<td data-value="33.33" class="pct low">33.33%</td>
-	<td data-value="24" class="abs low">8/24</td>
+	<td data-value="20.93" class="pct low">20.93%</td>
+	<td data-value="43" class="abs low">9/43</td>
 	</tr>
 
 <tr>
@@ -154,18 +154,33 @@
 	</tr>
 
 <tr>
-	<td class="file high" data-value="voiceInteractions.js"><a href="voiceInteractions.js.html">voiceInteractions.js</a></td>
-	<td data-value="86.95" class="pic high">
-	<div class="chart"><div class="cover-fill" style="width: 86%"></div><div class="cover-empty" style="width: 14%"></div></div>
+	<td class="file low" data-value="voiceInteractions.js"><a href="voiceInteractions.js.html">voiceInteractions.js</a></td>
+	<td data-value="13.04" class="pic low">
+	<div class="chart"><div class="cover-fill" style="width: 13%"></div><div class="cover-empty" style="width: 87%"></div></div>
 	</td>
-	<td data-value="86.95" class="pct high">86.95%</td>
-	<td data-value="69" class="abs high">60/69</td>
-	<td data-value="85.71" class="pct high">85.71%</td>
-	<td data-value="49" class="abs high">42/49</td>
+	<td data-value="13.04" class="pct low">13.04%</td>
+	<td data-value="69" class="abs low">9/69</td>
+	<td data-value="0" class="pct low">0%</td>
+	<td data-value="49" class="abs low">0/49</td>
+	<td data-value="0" class="pct low">0%</td>
+	<td data-value="7" class="abs low">0/7</td>
+	<td data-value="13.04" class="pct low">13.04%</td>
+	<td data-value="69" class="abs low">9/69</td>
+	</tr>
+
+<tr>
+	<td class="file high" data-value="voiceResponses.js"><a href="voiceResponses.js.html">voiceResponses.js</a></td>
+	<td data-value="100" class="pic high">
+	<div class="chart"><div class="cover-fill cover-full" style="width: 100%"></div><div class="cover-empty" style="width: 0%"></div></div>
+	</td>
 	<td data-value="100" class="pct high">100%</td>
-	<td data-value="7" class="abs high">7/7</td>
-	<td data-value="86.95" class="pct high">86.95%</td>
-	<td data-value="69" class="abs high">60/69</td>
+	<td data-value="11" class="abs high">11/11</td>
+	<td data-value="100" class="pct high">100%</td>
+	<td data-value="0" class="abs high">0/0</td>
+	<td data-value="100" class="pct high">100%</td>
+	<td data-value="0" class="abs high">0/0</td>
+	<td data-value="100" class="pct high">100%</td>
+	<td data-value="11" class="abs high">11/11</td>
 	</tr>
 
 </tbody>
@@ -176,7 +191,7 @@
             <div class='footer quiet pad2 space-top1 center small'>
                 Code coverage generated by
                 <a href="https://istanbul.js.org/" target="_blank" rel="noopener noreferrer">istanbul</a>
-                at 2025-10-11T01:53:09.221Z
+                at 2025-10-12T00:36:14.921Z
             </div>
         <script src="../../prettify.js"></script>
         <script>
