+++ backend-v1/tests/reports/coverage-html/backend-v1/services/index.html
@@ -23,30 +23,30 @@
         <div class='clearfix'>
             
             <div class='fl pad1y space-right2'>
-                <span class="strong">100% </span>
+                <span class="strong">10.92% </span>
                 <span class="quiet">Statements</span>
-                <span class='fraction'>61/61</span>
+                <span class='fraction'>13/119</span>
             </div>
         
             
             <div class='fl pad1y space-right2'>
-                <span class="strong">93.22% </span>
+                <span class="strong">9.3% </span>
                 <span class="quiet">Branches</span>
-                <span class='fraction'>55/59</span>
+                <span class='fraction'>8/86</span>
             </div>
         
             
             <div class='fl pad1y space-right2'>
-                <span class="strong">100% </span>
+                <span class="strong">10.52% </span>
                 <span class="quiet">Functions</span>
-                <span class='fraction'>9/9</span>
+                <span class='fraction'>2/19</span>
             </div>
         
             
             <div class='fl pad1y space-right2'>
-                <span class="strong">100% </span>
+                <span class="strong">11.11% </span>
                 <span class="quiet">Lines</span>
-                <span class='fraction'>60/60</span>
+                <span class='fraction'>13/117</span>
             </div>
         
             
@@ -61,7 +61,7 @@
             </div>
         </template>
     </div>
-    <div class='status-line high'></div>
+    <div class='status-line low'></div>
     <div class="pad1">
 <table class="coverage-summary">
 <thead>
@@ -79,18 +79,33 @@
 </tr>
 </thead>
 <tbody><tr>
-	<td class="file high" data-value="aiQuestionGenerator.js"><a href="aiQuestionGenerator.js.html">aiQuestionGenerator.js</a></td>
-	<td data-value="100" class="pic high">
-	<div class="chart"><div class="cover-fill cover-full" style="width: 100%"></div><div class="cover-empty" style="width: 0%"></div></div>
+	<td class="file low" data-value="aiQuestionGenerator.js"><a href="aiQuestionGenerator.js.html">aiQuestionGenerator.js</a></td>
+	<td data-value="13.11" class="pic low">
+	<div class="chart"><div class="cover-fill" style="width: 13%"></div><div class="cover-empty" style="width: 87%"></div></div>
 	</td>
-	<td data-value="100" class="pct high">100%</td>
-	<td data-value="61" class="abs high">61/61</td>
-	<td data-value="93.22" class="pct high">93.22%</td>
-	<td data-value="59" class="abs high">55/59</td>
-	<td data-value="100" class="pct high">100%</td>
-	<td data-value="9" class="abs high">9/9</td>
-	<td data-value="100" class="pct high">100%</td>
-	<td data-value="60" class="abs high">60/60</td>
+	<td data-value="13.11" class="pct low">13.11%</td>
+	<td data-value="61" class="abs low">8/61</td>
+	<td data-value="13.55" class="pct low">13.55%</td>
+	<td data-value="59" class="abs low">8/59</td>
+	<td data-value="11.11" class="pct low">11.11%</td>
+	<td data-value="9" class="abs low">1/9</td>
+	<td data-value="13.33" class="pct low">13.33%</td>
+	<td data-value="60" class="abs low">8/60</td>
+	</tr>
+
+<tr>
+	<td class="file low" data-value="assemblyAIService.js"><a href="assemblyAIService.js.html">assemblyAIService.js</a></td>
+	<td data-value="8.62" class="pic low">
+	<div class="chart"><div class="cover-fill" style="width: 8%"></div><div class="cover-empty" style="width: 92%"></div></div>
+	</td>
+	<td data-value="8.62" class="pct low">8.62%</td>
+	<td data-value="58" class="abs low">5/58</td>
+	<td data-value="0" class="pct low">0%</td>
+	<td data-value="27" class="abs low">0/27</td>
+	<td data-value="10" class="pct low">10%</td>
+	<td data-value="10" class="abs low">1/10</td>
+	<td data-value="8.77" class="pct low">8.77%</td>
+	<td data-value="57" class="abs low">5/57</td>
 	</tr>
 
 </tbody>
@@ -101,7 +116,7 @@
             <div class='footer quiet pad2 space-top1 center small'>
                 Code coverage generated by
                 <a href="https://istanbul.js.org/" target="_blank" rel="noopener noreferrer">istanbul</a>
-                at 2025-10-11T01:53:09.221Z
+                at 2025-10-12T00:36:14.921Z
             </div>
         <script src="../../prettify.js"></script>
         <script>
