+++ backend-v1/tests/reports/coverage-html/controllers/index.html
@@ -23,30 +23,30 @@
         <div class='clearfix'>
             
             <div class='fl pad1y space-right2'>
-                <span class="strong">78.06% </span>
+                <span class="strong">41.33% </span>
                 <span class="quiet">Statements</span>
-                <span class='fraction'>121/155</span>
+                <span class='fraction'>31/75</span>
             </div>
         
             
             <div class='fl pad1y space-right2'>
-                <span class="strong">55.84% </span>
+                <span class="strong">37.73% </span>
                 <span class="quiet">Branches</span>
-                <span class='fraction'>43/77</span>
+                <span class='fraction'>20/53</span>
             </div>
         
             
             <div class='fl pad1y space-right2'>
-                <span class="strong">90.47% </span>
+                <span class="strong">42.85% </span>
                 <span class="quiet">Functions</span>
-                <span class='fraction'>19/21</span>
+                <span class='fraction'>3/7</span>
             </div>
         
             
             <div class='fl pad1y space-right2'>
-                <span class="strong">77.77% </span>
+                <span class="strong">41.33% </span>
                 <span class="quiet">Lines</span>
-                <span class='fraction'>119/153</span>
+                <span class='fraction'>31/75</span>
             </div>
         
             
@@ -61,7 +61,7 @@
             </div>
         </template>
     </div>
-    <div class='status-line medium'></div>
+    <div class='status-line low'></div>
     <div class="pad1">
 <table class="coverage-summary">
 <thead>
@@ -79,63 +79,18 @@
 </tr>
 </thead>
 <tbody><tr>
-	<td class="file medium" data-value="aiController.js"><a href="aiController.js.html">aiController.js</a></td>
-	<td data-value="77.41" class="pic medium">
-	<div class="chart"><div class="cover-fill" style="width: 77%"></div><div class="cover-empty" style="width: 23%"></div></div>
+	<td class="file low" data-value="voiceController.js"><a href="voiceController.js.html">voiceController.js</a></td>
+	<td data-value="41.33" class="pic low">
+	<div class="chart"><div class="cover-fill" style="width: 41%"></div><div class="cover-empty" style="width: 59%"></div></div>
 	</td>
-	<td data-value="77.41" class="pct medium">77.41%</td>
-	<td data-value="31" class="abs medium">24/31</td>
-	<td data-value="69.56" class="pct medium">69.56%</td>
-	<td data-value="23" class="abs medium">16/23</td>
-	<td data-value="80" class="pct high">80%</td>
-	<td data-value="5" class="abs high">4/5</td>
-	<td data-value="77.41" class="pct medium">77.41%</td>
-	<td data-value="31" class="abs medium">24/31</td>
-	</tr>
-
-<tr>
-	<td class="file low" data-value="gamesController.js"><a href="gamesController.js.html">gamesController.js</a></td>
-	<td data-value="39.13" class="pic low">
-	<div class="chart"><div class="cover-fill" style="width: 39%"></div><div class="cover-empty" style="width: 61%"></div></div>
-	</td>
-	<td data-value="39.13" class="pct low">39.13%</td>
-	<td data-value="23" class="abs low">9/23</td>
-	<td data-value="0" class="pct low">0%</td>
-	<td data-value="11" class="abs low">0/11</td>
-	<td data-value="66.66" class="pct medium">66.66%</td>
-	<td data-value="3" class="abs medium">2/3</td>
-	<td data-value="36.36" class="pct low">36.36%</td>
-	<td data-value="22" class="abs low">8/22</td>
-	</tr>
-
-<tr>
-	<td class="file high" data-value="questionsController.js"><a href="questionsController.js.html">questionsController.js</a></td>
-	<td data-value="97.36" class="pic high">
-	<div class="chart"><div class="cover-fill" style="width: 97%"></div><div class="cover-empty" style="width: 3%"></div></div>
-	</td>
-	<td data-value="97.36" class="pct high">97.36%</td>
-	<td data-value="38" class="abs high">37/38</td>
-	<td data-value="75" class="pct medium">75%</td>
-	<td data-value="4" class="abs medium">3/4</td>
-	<td data-value="100" class="pct high">100%</td>
-	<td data-value="7" class="abs high">7/7</td>
-	<td data-value="97.29" class="pct high">97.29%</td>
-	<td data-value="37" class="abs high">36/37</td>
-	</tr>
-
-<tr>
-	<td class="file high" data-value="usersController.js"><a href="usersController.js.html">usersController.js</a></td>
-	<td data-value="80.95" class="pic high">
-	<div class="chart"><div class="cover-fill" style="width: 80%"></div><div class="cover-empty" style="width: 20%"></div></div>
-	</td>
-	<td data-value="80.95" class="pct high">80.95%</td>
-	<td data-value="63" class="abs high">51/63</td>
-	<td data-value="61.53" class="pct medium">61.53%</td>
-	<td data-value="39" class="abs medium">24/39</td>
-	<td data-value="100" class="pct high">100%</td>
-	<td data-value="6" class="abs high">6/6</td>
-	<td data-value="80.95" class="pct high">80.95%</td>
-	<td data-value="63" class="abs high">51/63</td>
+	<td data-value="41.33" class="pct low">41.33%</td>
+	<td data-value="75" class="abs low">31/75</td>
+	<td data-value="37.73" class="pct low">37.73%</td>
+	<td data-value="53" class="abs low">20/53</td>
+	<td data-value="42.85" class="pct low">42.85%</td>
+	<td data-value="7" class="abs low">3/7</td>
+	<td data-value="41.33" class="pct low">41.33%</td>
+	<td data-value="75" class="abs low">31/75</td>
 	</tr>
 
 </tbody>
@@ -146,7 +101,7 @@
             <div class='footer quiet pad2 space-top1 center small'>
                 Code coverage generated by
                 <a href="https://istanbul.js.org/" target="_blank" rel="noopener noreferrer">istanbul</a>
-                at 2025-10-10T20:38:42.555Z
+                at 2025-10-12T00:39:19.003Z
             </div>
         <script src="../prettify.js"></script>
         <script>
