+++ backend-v1/tests/reports/coverage-html/coverage-summary.json
@@ -1,17 +1,8 @@
-{"total": {"lines":{"total":589,"covered":349,"skipped":0,"pct":59.25},"statements":{"total":603,"covered":352,"skipped":0,"pct":58.37},"functions":{"total":67,"covered":37,"skipped":0,"pct":55.22},"branches":{"total":341,"covered":159,"skipped":0,"pct":46.62},"branchesTrue":{"total":0,"covered":0,"skipped":0,"pct":100}}
-,"/home/ervin/Desktop/PROYECTO-2/backend-v1/firebase.js": {"lines":{"total":9,"covered":9,"skipped":0,"pct":100},"functions":{"total":0,"covered":0,"skipped":0,"pct":100},"statements":{"total":9,"covered":9,"skipped":0,"pct":100},"branches":{"total":2,"covered":2,"skipped":0,"pct":100}}
-,"/home/ervin/Desktop/PROYECTO-2/backend-v1/hybridServer.js": {"lines":{"total":215,"covered":26,"skipped":0,"pct":12.09},"functions":{"total":21,"covered":1,"skipped":0,"pct":4.76},"statements":{"total":226,"covered":26,"skipped":0,"pct":11.5},"branches":{"total":148,"covered":1,"skipped":0,"pct":0.67}}
-,"/home/ervin/Desktop/PROYECTO-2/backend-v1/controllers/aiController.js": {"lines":{"total":31,"covered":24,"skipped":0,"pct":77.41},"functions":{"total":5,"covered":4,"skipped":0,"pct":80},"statements":{"total":31,"covered":24,"skipped":0,"pct":77.41},"branches":{"total":23,"covered":16,"skipped":0,"pct":69.56}}
-,"/home/ervin/Desktop/PROYECTO-2/backend-v1/controllers/gamesController.js": {"lines":{"total":22,"covered":8,"skipped":0,"pct":36.36},"functions":{"total":3,"covered":2,"skipped":0,"pct":66.66},"statements":{"total":23,"covered":9,"skipped":0,"pct":39.13},"branches":{"total":11,"covered":0,"skipped":0,"pct":0}}
-,"/home/ervin/Desktop/PROYECTO-2/backend-v1/controllers/questionsController.js": {"lines":{"total":37,"covered":37,"skipped":0,"pct":100},"functions":{"total":7,"covered":7,"skipped":0,"pct":100},"statements":{"total":38,"covered":38,"skipped":0,"pct":100},"branches":{"total":4,"covered":4,"skipped":0,"pct":100}}
-,"/home/ervin/Desktop/PROYECTO-2/backend-v1/controllers/usersController.js": {"lines":{"total":63,"covered":62,"skipped":0,"pct":98.41},"functions":{"total":6,"covered":6,"skipped":0,"pct":100},"statements":{"total":63,"covered":62,"skipped":0,"pct":98.41},"branches":{"total":39,"covered":35,"skipped":0,"pct":89.74}}
-,"/home/ervin/Desktop/PROYECTO-2/backend-v1/middleware/authenticate.js": {"lines":{"total":11,"covered":11,"skipped":0,"pct":100},"functions":{"total":1,"covered":1,"skipped":0,"pct":100},"statements":{"total":11,"covered":11,"skipped":0,"pct":100},"branches":{"total":4,"covered":4,"skipped":0,"pct":100}}
-,"/home/ervin/Desktop/PROYECTO-2/backend-v1/middleware/rateLimiter.js": {"lines":{"total":6,"covered":6,"skipped":0,"pct":100},"functions":{"total":0,"covered":0,"skipped":0,"pct":100},"statements":{"total":6,"covered":6,"skipped":0,"pct":100},"branches":{"total":0,"covered":0,"skipped":0,"pct":100}}
-,"/home/ervin/Desktop/PROYECTO-2/backend-v1/routes/adminAccessibility.js": {"lines":{"total":24,"covered":8,"skipped":0,"pct":33.33},"functions":{"total":4,"covered":0,"skipped":0,"pct":0},"statements":{"total":24,"covered":8,"skipped":0,"pct":33.33},"branches":{"total":2,"covered":0,"skipped":0,"pct":0}}
-,"/home/ervin/Desktop/PROYECTO-2/backend-v1/routes/ai.js": {"lines":{"total":13,"covered":9,"skipped":0,"pct":69.23},"functions":{"total":4,"covered":0,"skipped":0,"pct":0},"statements":{"total":13,"covered":9,"skipped":0,"pct":69.23},"branches":{"total":0,"covered":0,"skipped":0,"pct":100}}
-,"/home/ervin/Desktop/PROYECTO-2/backend-v1/routes/games.js": {"lines":{"total":7,"covered":7,"skipped":0,"pct":100},"functions":{"total":0,"covered":0,"skipped":0,"pct":100},"statements":{"total":7,"covered":7,"skipped":0,"pct":100},"branches":{"total":0,"covered":0,"skipped":0,"pct":100}}
-,"/home/ervin/Desktop/PROYECTO-2/backend-v1/routes/questions.js": {"lines":{"total":10,"covered":10,"skipped":0,"pct":100},"functions":{"total":0,"covered":0,"skipped":0,"pct":100},"statements":{"total":10,"covered":10,"skipped":0,"pct":100},"branches":{"total":0,"covered":0,"skipped":0,"pct":100}}
-,"/home/ervin/Desktop/PROYECTO-2/backend-v1/routes/users.js": {"lines":{"total":12,"covered":12,"skipped":0,"pct":100},"functions":{"total":0,"covered":0,"skipped":0,"pct":100},"statements":{"total":12,"covered":12,"skipped":0,"pct":100},"branches":{"total":0,"covered":0,"skipped":0,"pct":100}}
-,"/home/ervin/Desktop/PROYECTO-2/backend-v1/routes/voiceInteractions.js": {"lines":{"total":69,"covered":60,"skipped":0,"pct":86.95},"functions":{"total":7,"covered":7,"skipped":0,"pct":100},"statements":{"total":69,"covered":60,"skipped":0,"pct":86.95},"branches":{"total":49,"covered":42,"skipped":0,"pct":85.71}}
-,"/home/ervin/Desktop/PROYECTO-2/backend-v1/services/aiQuestionGenerator.js": {"lines":{"total":60,"covered":60,"skipped":0,"pct":100},"functions":{"total":9,"covered":9,"skipped":0,"pct":100},"statements":{"total":61,"covered":61,"skipped":0,"pct":100},"branches":{"total":59,"covered":55,"skipped":0,"pct":93.22}}
+{"total": {"lines":{"total":206,"covered":68,"skipped":0,"pct":33},"statements":{"total":210,"covered":70,"skipped":0,"pct":33.33},"functions":{"total":29,"covered":8,"skipped":0,"pct":27.58},"branches":{"total":110,"covered":27,"skipped":0,"pct":24.54},"branchesTrue":{"total":0,"covered":0,"skipped":0,"pct":100}}
+,"/workspace/backend-v1/controllers/voiceController.js": {"lines":{"total":75,"covered":31,"skipped":0,"pct":41.33},"functions":{"total":7,"covered":3,"skipped":0,"pct":42.85},"statements":{"total":75,"covered":31,"skipped":0,"pct":41.33},"branches":{"total":53,"covered":20,"skipped":0,"pct":37.73}}
+,"/workspace/backend-v1/middleware/authenticate.js": {"lines":{"total":11,"covered":2,"skipped":0,"pct":18.18},"functions":{"total":1,"covered":0,"skipped":0,"pct":0},"statements":{"total":11,"covered":2,"skipped":0,"pct":18.18},"branches":{"total":4,"covered":0,"skipped":0,"pct":0}}
+,"/workspace/backend-v1/middleware/rateLimiter.js": {"lines":{"total":6,"covered":6,"skipped":0,"pct":100},"functions":{"total":0,"covered":0,"skipped":0,"pct":100},"statements":{"total":6,"covered":6,"skipped":0,"pct":100},"branches":{"total":0,"covered":0,"skipped":0,"pct":100}}
+,"/workspace/backend-v1/routes/voiceResponses.js": {"lines":{"total":11,"covered":11,"skipped":0,"pct":100},"functions":{"total":0,"covered":0,"skipped":0,"pct":100},"statements":{"total":11,"covered":11,"skipped":0,"pct":100},"branches":{"total":0,"covered":0,"skipped":0,"pct":100}}
+,"/workspace/backend-v1/services/assemblyAIService.js": {"lines":{"total":57,"covered":5,"skipped":0,"pct":8.77},"functions":{"total":10,"covered":1,"skipped":0,"pct":10},"statements":{"total":58,"covered":5,"skipped":0,"pct":8.62},"branches":{"total":27,"covered":0,"skipped":0,"pct":0}}
+,"/workspace/backend-v1/utils/voiceRecognition.js": {"lines":{"total":46,"covered":13,"skipped":0,"pct":28.26},"functions":{"total":11,"covered":4,"skipped":0,"pct":36.36},"statements":{"total":49,"covered":15,"skipped":0,"pct":30.61},"branches":{"total":26,"covered":7,"skipped":0,"pct":26.92}}
 }
