+++ backend-v1/tests/reports/coverage-html/middleware/authenticate.js.html
@@ -23,30 +23,30 @@
         <div class='clearfix'>
             
             <div class='fl pad1y space-right2'>
-                <span class="strong">100% </span>
+                <span class="strong">18.18% </span>
                 <span class="quiet">Statements</span>
-                <span class='fraction'>11/11</span>
+                <span class='fraction'>2/11</span>
             </div>
         
             
             <div class='fl pad1y space-right2'>
-                <span class="strong">100% </span>
+                <span class="strong">0% </span>
                 <span class="quiet">Branches</span>
-                <span class='fraction'>4/4</span>
+                <span class='fraction'>0/4</span>
             </div>
         
             
             <div class='fl pad1y space-right2'>
-                <span class="strong">100% </span>
+                <span class="strong">0% </span>
                 <span class="quiet">Functions</span>
-                <span class='fraction'>1/1</span>
+                <span class='fraction'>0/1</span>
             </div>
         
             
             <div class='fl pad1y space-right2'>
-                <span class="strong">100% </span>
+                <span class="strong">18.18% </span>
                 <span class="quiet">Lines</span>
-                <span class='fraction'>11/11</span>
+                <span class='fraction'>2/11</span>
             </div>
         
             
@@ -61,7 +61,7 @@
             </div>
         </template>
     </div>
-    <div class='status-line high'></div>
+    <div class='status-line low'></div>
     <pre><table class="coverage">
 <tr><td class="line-count quiet"><a name='L1'></a><a href='#L1'>1</a>
 <a name='L2'></a><a href='#L2'>2</a>
@@ -84,34 +84,34 @@
 <span class="cline-any cline-neutral">&nbsp;</span>
 <span class="cline-any cline-neutral">&nbsp;</span>
 <span class="cline-any cline-yes">1x</span>
-<span class="cline-any cline-yes">3x</span>
-<span class="cline-any cline-yes">3x</span>
-<span class="cline-any cline-yes">1x</span>
+<span class="cline-any cline-no">&nbsp;</span>
+<span class="cline-any cline-no">&nbsp;</span>
+<span class="cline-any cline-no">&nbsp;</span>
 <span class="cline-any cline-neutral">&nbsp;</span>
-<span class="cline-any cline-yes">2x</span>
-<span class="cline-any cline-yes">2x</span>
-<span class="cline-any cline-yes">2x</span>
-<span class="cline-any cline-yes">1x</span>
-<span class="cline-any cline-yes">1x</span>
+<span class="cline-any cline-no">&nbsp;</span>
+<span class="cline-any cline-no">&nbsp;</span>
+<span class="cline-any cline-no">&nbsp;</span>
+<span class="cline-any cline-no">&nbsp;</span>
+<span class="cline-any cline-no">&nbsp;</span>
 <span class="cline-any cline-neutral">&nbsp;</span>
-<span class="cline-any cline-yes">1x</span>
+<span class="cline-any cline-no">&nbsp;</span>
 <span class="cline-any cline-neutral">&nbsp;</span>
 <span class="cline-any cline-neutral">&nbsp;</span>
 <span class="cline-any cline-neutral">&nbsp;</span></td><td class="text"><pre class="prettyprint lang-js">const { auth } = require('../firebase');
 &nbsp;
 // Middleware para autenticar peticiones usando el token JWT de Firebase
-module.exports = async function authenticate(req, res, next) {
-  const authHeader = req.headers.authorization;
-  if (!authHeader || !authHeader.startsWith('Bearer ')) {
-    return res.status(401).json({ error: 'No token provided' });
+module.exports = async function <span class="fstat-no" title="function not covered" >authenticate(</span>req, res, next) {
+  const authHeader = <span class="cstat-no" title="statement not covered" >req.headers.authorization;</span>
+<span class="cstat-no" title="statement not covered" >  if (!authHeader || !authHeader.startsWith('Bearer ')) {</span>
+<span class="cstat-no" title="statement not covered" >    return res.status(401).json({ error: 'No token provided' });</span>
   }
-  const token = authHeader.split(' ')[1];
-  try {
-    const decoded = await auth.verifyIdToken(token);
-    req.user = decoded;
-    next();
+  const token = <span class="cstat-no" title="statement not covered" >authHeader.split(' ')[1];</span>
+<span class="cstat-no" title="statement not covered" >  try {</span>
+    const decoded = <span class="cstat-no" title="statement not covered" >await auth.verifyIdToken(token);</span>
+<span class="cstat-no" title="statement not covered" >    req.user = decoded;</span>
+<span class="cstat-no" title="statement not covered" >    next();</span>
   } catch (err) {
-    return res.status(401).json({ error: 'Invalid or expired token' });
+<span class="cstat-no" title="statement not covered" >    return res.status(401).json({ error: 'Invalid or expired token' });</span>
   }
 };
 &nbsp;</pre></td></tr></table></pre>
@@ -121,7 +121,7 @@ module.exports = async function authenticate(req, res, next) {
             <div class='footer quiet pad2 space-top1 center small'>
                 Code coverage generated by
                 <a href="https://istanbul.js.org/" target="_blank" rel="noopener noreferrer">istanbul</a>
-                at 2025-10-10T20:38:42.555Z
+                at 2025-10-12T00:39:19.003Z
             </div>
         <script src="../prettify.js"></script>
         <script>
