+++ backend-v1/tests/reports/coverage-html/routes/voiceResponses.js.html
@@ -3,15 +3,15 @@
 <html lang="en">
 
 <head>
-    <title>Code coverage report for backend-v1/routes/users.js</title>
+    <title>Code coverage report for routes/voiceResponses.js</title>
     <meta charset="utf-8" />
-    <link rel="stylesheet" href="../../prettify.css" />
-    <link rel="stylesheet" href="../../base.css" />
-    <link rel="shortcut icon" type="image/x-icon" href="../../favicon.png" />
+    <link rel="stylesheet" href="../prettify.css" />
+    <link rel="stylesheet" href="../base.css" />
+    <link rel="shortcut icon" type="image/x-icon" href="../favicon.png" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <style type='text/css'>
         .coverage-summary .sorter {
-            background-image: url(../../sort-arrow-sprite.png);
+            background-image: url(../sort-arrow-sprite.png);
         }
     </style>
 </head>
@@ -19,13 +19,13 @@
 <body>
 <div class='wrapper'>
     <div class='pad1'>
-        <h1><a href="../../index.html">All files</a> / <a href="index.html">backend-v1/routes</a> users.js</h1>
+        <h1><a href="../index.html">All files</a> / <a href="index.html">routes</a> voiceResponses.js</h1>
         <div class='clearfix'>
             
             <div class='fl pad1y space-right2'>
                 <span class="strong">100% </span>
                 <span class="quiet">Statements</span>
-                <span class='fraction'>12/12</span>
+                <span class='fraction'>11/11</span>
             </div>
         
             
@@ -46,7 +46,7 @@
             <div class='fl pad1y space-right2'>
                 <span class="strong">100% </span>
                 <span class="quiet">Lines</span>
-                <span class='fraction'>12/12</span>
+                <span class='fraction'>11/11</span>
             </div>
         
             
@@ -86,63 +86,63 @@
 <a name='L21'></a><a href='#L21'>21</a>
 <a name='L22'></a><a href='#L22'>22</a></td><td class="line-coverage quiet"><span class="cline-any cline-yes">1x</span>
 <span class="cline-any cline-yes">1x</span>
-<span class="cline-any cline-neutral">&nbsp;</span>
 <span class="cline-any cline-yes">1x</span>
 <span class="cline-any cline-yes">1x</span>
-<span class="cline-any cline-neutral">&nbsp;</span>
-<span class="cline-any cline-neutral">&nbsp;</span>
-<span class="cline-any cline-neutral">&nbsp;</span>
+<span class="cline-any cline-yes">1x</span>
 <span class="cline-any cline-neutral">&nbsp;</span>
 <span class="cline-any cline-neutral">&nbsp;</span>
 <span class="cline-any cline-yes">1x</span>
 <span class="cline-any cline-neutral">&nbsp;</span>
 <span class="cline-any cline-neutral">&nbsp;</span>
 <span class="cline-any cline-yes">1x</span>
-<span class="cline-any cline-yes">1x</span>
-<span class="cline-any cline-yes">1x</span>
-<span class="cline-any cline-yes">1x</span>
-<span class="cline-any cline-yes">1x</span>
-<span class="cline-any cline-yes">1x</span>
+<span class="cline-any cline-neutral">&nbsp;</span>
 <span class="cline-any cline-neutral">&nbsp;</span>
 <span class="cline-any cline-yes">1x</span>
-<span class="cline-any cline-neutral">&nbsp;</span></td><td class="text"><pre class="prettyprint lang-js">const express = require('express');
+<span class="cline-any cline-neutral">&nbsp;</span>
+<span class="cline-any cline-neutral">&nbsp;</span>
+<span class="cline-any cline-yes">1x</span>
+<span class="cline-any cline-neutral">&nbsp;</span>
+<span class="cline-any cline-neutral">&nbsp;</span>
+<span class="cline-any cline-yes">1x</span>
+<span class="cline-any cline-neutral">&nbsp;</span>
+<span class="cline-any cline-yes">1x</span></td><td class="text"><pre class="prettyprint lang-js">const express = require('express');
 const router = express.Router();
-&nbsp;
-const usersController = require('../controllers/usersController');
+const voiceController = require('../controllers/voiceController');
 const authenticate = require('../middleware/authenticate');
-const { 
-  registerLimiter, 
-  passwordRecoveryLimiter, 
-  profileUpdateLimiter, 
-  generalUserLimiter 
-} = require('../middleware/rateLimiter');
+const { generalUserLimiter } = require('../middleware/rateLimiter');
 &nbsp;
-// Aplicar rate limiting específico para cada endpoint
-router.post('/register', registerLimiter, usersController.register);
-router.post('/login', generalUserLimiter, usersController.login); 
-router.post('/recover-password', passwordRecoveryLimiter, usersController.recoverPassword);
-router.put('/me/profile', profileUpdateLimiter, authenticate, usersController.updateProfile);
-router.get('/me/stats', generalUserLimiter, authenticate, usersController.getStats);
-router.get('/me/history', generalUserLimiter, authenticate, usersController.getHistory);
+// [HU8] Validar respuesta de voz
+router.post('/validate', generalUserLimiter, voiceController.validateVoiceResponse);
 &nbsp;
-module.exports = router;
-&nbsp;</pre></td></tr></table></pre>
+// [HU8] Procesar respuesta de voz
+router.post('/process', generalUserLimiter, voiceController.processVoiceResponse);
+&nbsp;
+// [HU8] Obtener estadísticas de reconocimiento de voz
+router.get('/stats/:userId', generalUserLimiter, voiceController.getVoiceRecognitionStats);
+&nbsp;
+// [HU8] Procesar audio con AssemblyAI
+router.post('/process-audio', generalUserLimiter, voiceController.processAudioWithAssemblyAI);
+&nbsp;
+// [HU8] Verificar estado de AssemblyAI
+router.get('/assemblyai/status', generalUserLimiter, voiceController.checkAssemblyAIStatus);
+&nbsp;
+module.exports = router;</pre></td></tr></table></pre>
 
                 <div class='push'></div><!-- for sticky footer -->
             </div><!-- /wrapper -->
             <div class='footer quiet pad2 space-top1 center small'>
                 Code coverage generated by
                 <a href="https://istanbul.js.org/" target="_blank" rel="noopener noreferrer">istanbul</a>
-                at 2025-10-12T00:36:14.921Z
+                at 2025-10-12T00:39:19.003Z
             </div>
-        <script src="../../prettify.js"></script>
+        <script src="../prettify.js"></script>
         <script>
             window.onload = function () {
                 prettyPrint();
             };
         </script>
-        <script src="../../sorter.js"></script>
-        <script src="../../block-navigation.js"></script>
+        <script src="../sorter.js"></script>
+        <script src="../block-navigation.js"></script>
     </body>
 </html>
