+++ backend-v1/tests/reports/coverage-html/utils/index.html
@@ -3,7 +3,7 @@
 <html lang="en">
 
 <head>
-    <title>Code coverage report for services</title>
+    <title>Code coverage report for utils</title>
     <meta charset="utf-8" />
     <link rel="stylesheet" href="../prettify.css" />
     <link rel="stylesheet" href="../base.css" />
@@ -19,34 +19,34 @@
 <body>
 <div class='wrapper'>
     <div class='pad1'>
-        <h1><a href="../index.html">All files</a> services</h1>
+        <h1><a href="../index.html">All files</a> utils</h1>
         <div class='clearfix'>
             
             <div class='fl pad1y space-right2'>
-                <span class="strong">8.62% </span>
+                <span class="strong">30.61% </span>
                 <span class="quiet">Statements</span>
-                <span class='fraction'>5/58</span>
+                <span class='fraction'>15/49</span>
             </div>
         
             
             <div class='fl pad1y space-right2'>
-                <span class="strong">0% </span>
+                <span class="strong">26.92% </span>
                 <span class="quiet">Branches</span>
-                <span class='fraction'>0/27</span>
+                <span class='fraction'>7/26</span>
             </div>
         
             
             <div class='fl pad1y space-right2'>
-                <span class="strong">10% </span>
+                <span class="strong">36.36% </span>
                 <span class="quiet">Functions</span>
-                <span class='fraction'>1/10</span>
+                <span class='fraction'>4/11</span>
             </div>
         
             
             <div class='fl pad1y space-right2'>
-                <span class="strong">8.77% </span>
+                <span class="strong">28.26% </span>
                 <span class="quiet">Lines</span>
-                <span class='fraction'>5/57</span>
+                <span class='fraction'>13/46</span>
             </div>
         
             
@@ -79,18 +79,18 @@
 </tr>
 </thead>
 <tbody><tr>
-	<td class="file low" data-value="assemblyAIService.js"><a href="assemblyAIService.js.html">assemblyAIService.js</a></td>
-	<td data-value="8.62" class="pic low">
-	<div class="chart"><div class="cover-fill" style="width: 8%"></div><div class="cover-empty" style="width: 92%"></div></div>
+	<td class="file low" data-value="voiceRecognition.js"><a href="voiceRecognition.js.html">voiceRecognition.js</a></td>
+	<td data-value="30.61" class="pic low">
+	<div class="chart"><div class="cover-fill" style="width: 30%"></div><div class="cover-empty" style="width: 70%"></div></div>
 	</td>
-	<td data-value="8.62" class="pct low">8.62%</td>
-	<td data-value="58" class="abs low">5/58</td>
-	<td data-value="0" class="pct low">0%</td>
-	<td data-value="27" class="abs low">0/27</td>
-	<td data-value="10" class="pct low">10%</td>
-	<td data-value="10" class="abs low">1/10</td>
-	<td data-value="8.77" class="pct low">8.77%</td>
-	<td data-value="57" class="abs low">5/57</td>
+	<td data-value="30.61" class="pct low">30.61%</td>
+	<td data-value="49" class="abs low">15/49</td>
+	<td data-value="26.92" class="pct low">26.92%</td>
+	<td data-value="26" class="abs low">7/26</td>
+	<td data-value="36.36" class="pct low">36.36%</td>
+	<td data-value="11" class="abs low">4/11</td>
+	<td data-value="28.26" class="pct low">28.26%</td>
+	<td data-value="46" class="abs low">13/46</td>
 	</tr>
 
 </tbody>
